panel/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
passlib==1.7.4
bcrypt==4.1.2
psutil==5.9.8
docker==7.0.0
PyYAML==6.0.1
